@@ -1,5 +1,5 @@
 {
-  "name": "template-project",
+  "name": "mfs-client",
   "private": true,
   "version": "0.0.0",
   "type": "module",
@@ -10,39 +10,26 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@radix-ui/react-dropdown-menu": "^2.1.1",
-    "@radix-ui/react-slot": "^1.0.2",
-    "@tanstack/react-query": "^5.35.1",
-    "axios": "^1.6.8",
-    "class-variance-authority": "^0.7.0",
-    "clsx": "^2.1.1",
-    "firebase": "^10.10.0",
+    "dotenv": "^16.4.5",
     "localforage": "^1.10.0",
-    "lucide-react": "^0.379.0",
     "match-sorter": "^6.3.4",
-    "moment": "^2.30.1",
-    "prop-types": "^15.8.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-hot-toast": "^2.4.1",
-    "react-icons": "^5.2.1",
-    "react-router-dom": "^6.22.3",
-    "react-spinners": "^0.13.8",
-    "sort-by": "^1.2.0",
-    "tailwind-merge": "^2.3.0",
-    "tailwindcss-animate": "^1.0.7"
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.24.1",
+    "sort-by": "^1.2.0"
   },
   "devDependencies": {
-    "@types/react": "^18.2.66",
-    "@types/react-dom": "^18.2.22",
-    "@vitejs/plugin-react": "^4.2.1",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@vitejs/plugin-react": "^4.3.1",
     "autoprefixer": "^10.4.19",
+    "daisyui": "^4.12.10",
     "eslint": "^8.57.0",
-    "eslint-plugin-react": "^7.34.1",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.6",
-    "postcss": "^8.4.38",
-    "tailwindcss": "^3.4.3",
-    "vite": "^5.2.11"
+    "eslint-plugin-react": "^7.34.2",
+    "eslint-plugin-react-hooks": "^4.6.2",
+    "eslint-plugin-react-refresh": "^0.4.7",
+    "postcss": "^8.4.39",
+    "tailwindcss": "^3.4.5",
+    "vite": "^5.3.1"
   }
 }
