@@ -1,6 +1,6 @@
 module.exports = {
   root: true,
-  env: { browser: true, es2020: true,node:true },
+  env: { browser: true, node: true, es2020: true },
   extends: [
     'eslint:recommended',
     'plugin:react/recommended',
@@ -12,7 +12,6 @@ module.exports = {
   settings: { react: { version: '18.2' } },
   plugins: ['react-refresh'],
   rules: {
-    'react/jsx-uses-react': 'warn',
     'react/jsx-no-target-blank': 'off',
     'react-refresh/only-export-components': [
       'warn',
